fix(pkg): key the cache on the PHP grammar, and make the PHP modules a tree
Two promotion blockers from the review of #334.

- `persistence._GRAMMAR_MODULES` named every grammar except `tree_sitter_php`, so
  the extractor fingerprint did not change when the PHP extra was installed and a
  graph cached before it — with no PHP facts — was served after it for as long as
  the commit stood still. Listed now; a test checks the list against
  `doctor.EXTRA_PROBES`, the one place every extra is enumerated, so the next
  grammar cannot be forgotten the same way.
- `php_routes.py` and `php_orm.py` imported the name resolver and `_TypeRec` from
  `php_extractor.py`, which imports them back (CodeQL py/cyclic-import, six
  alerts). It never failed at import time because the extractor's side was
  function-local, but a cycle that works because of where an import sits is one
  refactor from an ImportError. The shared pieces now live in `php_names.py`, a
  leaf that imports nothing from its siblings — the shape go_routes.py has always
  had. No behaviour change: same names, same ids, same facts.

STATE-OF-SPINE's gated source-module and test-function counts bumped to match.
