+++ src/orchestrator/pkg/persistence.py
@@ -168,6 +168,7 @@ _GRAMMAR_MODULES = (
     "tree_sitter_c",
     "tree_sitter_cpp",
     "tree_sitter_go",
+    "tree_sitter_php",
     "sqlglot",
 )
