+++ src/orchestrator/pkg/php_extractor.py
@@ -42,12 +42,23 @@ PHP source.
 from __future__ import annotations
 
 import re
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from pathlib import Path, PurePosixPath
 from typing import TYPE_CHECKING, Any
 
 from orchestrator.pkg.extractor import rel_module_name
 from orchestrator.pkg.facts import Edge, EdgeKind, FactBatch, Node, NodeKind, Provenance
+
+# The name resolver and the collected-type record are shared with php_routes.py and
+# php_orm.py, so they live in a leaf module neither side imports the other for.
+from orchestrator.pkg.php_names import (
+    _join,
+    _resolve_type_name,
+    _text,
+    _to_dotted,
+    _TypeRec,
+    _use_declaration_targets,
+)
 
 if TYPE_CHECKING:
     from tree_sitter import Node as TSNode
@@ -77,42 +88,6 @@ _REQUIRE_KINDS = frozenset(
         "include_once_expression",
     }
 )
-
-
-def _to_dotted(name: str) -> str:
-    """``App\\Http\\Controllers`` -> ``App.Http.Controllers`` (D2)."""
-    return name.replace("\\", ".").strip(".")
-
-
-def _join(namespace: str, name: str) -> str:
-    if namespace and name:
-        return f"{namespace}.{name}"
-    return name or namespace
-
-
-@dataclass
-class _TypeRec:
-    """A collected class/interface/trait/enum — the working set the CALLS pass (P2)
-    reasons over, once every declaration in the file is known."""
-
-    type_id: str
-    name: str
-    namespace: str
-    use_map: dict[str, str]
-    use_func_map: dict[str, str]
-    node: TSNode  # the declaration node itself — php_routes.py reads its attributes (P3)
-    # The single `extends` target's raw name (classes only — D3's trait/interface
-    # `use`/`implements` never feed `parent::`), for CALLS row 3.
-    base_name: str | None = None
-    methods: list[tuple[str, str, TSNode]] = field(default_factory=list)  # (name, id, node)
-    # Trait names this type `use`s (D3) — resolved lazily in the CALLS pass so a
-    # same-file trait's methods count as this type's own for `$this->`/`self::` calls
-    # it doesn't override (the `traits` corpus case's whole point).
-    trait_names: list[str] = field(default_factory=list)
-    # name -> raw type text, for every TYPED property and TYPED promoted ctor param
-    # (both become properties reachable via `$this->prop->m()`) — P3's typed-receiver
-    # rule (§3.2 rows 7-8). An untyped property/param is simply absent here.
-    typed_props: dict[str, str] = field(default_factory=dict)
 
 
 @dataclass
@@ -797,76 +772,6 @@ def _emit_receiver_call(
         batch.add_edge(Edge(caller_id, target, EdgeKind.CALLS, Provenance(rel, line)))
 
 
-# --- use-declaration parsing -------------------------------------------------
-
-
-def _use_clause(clause: TSNode, source: bytes) -> tuple[str, str, str]:
-    """``(fqn, bind_name, kind)`` for one ``namespace_use_clause``.
-
-    ``kind`` is ``class`` unless a ``function``/``const`` marker is present. ``fqn`` is
-    left backslash-separated (the caller joins group prefixes before converting to dots
-    once). ``bind_name`` is the alias if aliased, else the FQN's own last segment.
-    """
-    kind = "class"
-    type_tok = clause.child_by_field_name("type")
-    if type_tok is not None and _text(type_tok, source) in ("function", "const"):
-        kind = _text(type_tok, source)
-    name_node = next((c for c in clause.named_children if c.type in ("qualified_name", "name")), None)
-    fqn = _text(name_node, source) if name_node is not None else ""
-    alias_node = clause.child_by_field_name("alias")
-    bind = _text(alias_node, source) if alias_node is not None else fqn.rsplit("\\", 1)[-1]
-    return fqn, bind, kind
-
-
-def _use_declaration_targets(decl: TSNode, source: bytes) -> list[tuple[str, str, str]]:
-    """``(fqn_dotted, bind_name, kind)`` for every clause in a ``namespace_use_declaration``,
-    grouped (``use App\\Models\\{Customer, Invoice as Inv};``) or not."""
-    out: list[tuple[str, str, str]] = []
-    prefix = ""
-    group: TSNode | None = None
-    for child in decl.named_children:
-        if child.type == "namespace_name":
-            prefix = _text(child, source)
-        elif child.type == "namespace_use_group":
-            group = child
-        elif child.type == "namespace_use_clause":
-            fqn, bind, kind = _use_clause(child, source)
-            out.append((_to_dotted(fqn), bind, kind))
-    if group is not None:
-        for clause in group.named_children:
-            if clause.type != "namespace_use_clause":
-                continue
-            fqn, bind, kind = _use_clause(clause, source)
-            full = f"{prefix}\\{fqn}" if prefix else fqn
-            out.append((_to_dotted(full), bind, kind))
-    return out
-
-
-# --- PHP's class-name resolution (D0/§2) -------------------------------------
-
-
-def _resolve_type_name(name: str, namespace: str, use_map: dict[str, str]) -> tuple[str, bool]:
-    """Fully-qualified, qualified, or unqualified -> current namespace unless `use`d
-    (the PHP RFC rule). Returns ``(dotted_id, is_guess)``: ``is_guess`` is True only
-    for the same-namespace assumption made when nothing resolves the name — the one
-    case `finalize` may need to correct. Every other shape (a leading `\\`, or a hit
-    in the `use` map) is a claim read directly off the source and must never be
-    treated as provisional, even though it is very often *external* — see the
-    `finalize` docstring for why the distinction matters."""
-    name = name.strip()
-    if name.startswith("\\"):
-        return _to_dotted(name), False  # fully qualified — already absolute
-    if "\\" in name:  # qualified: resolve the first segment, keep the rest
-        head, _, rest = name.partition("\\")
-        if head in use_map:
-            base = use_map[head]
-            return (f"{base}.{_to_dotted(rest)}" if rest else base), False
-        return _join(namespace, _to_dotted(name)), True
-    if name in use_map:  # unqualified, explicitly `use`d
-        return use_map[name], False
-    return _join(namespace, name), True  # unqualified, not `use`d — assume current namespace
-
-
 # --- require/include literal detection (D7) ----------------------------------
 
 
@@ -1046,12 +951,6 @@ def _string_content(node: TSNode, source: bytes) -> str:
 def _field_text(node: TSNode, field: str, source: bytes) -> str:
     child = node.child_by_field_name(field)
     return _text(child, source) if child is not None else ""
-
-
-def _text(node: TSNode | None, source: bytes) -> str:
-    if node is None:
-        return ""
-    return source[node.start_byte : node.end_byte].decode("utf-8", "replace").strip()
 
 
 def _php_parser() -> Any:
